travel_project/requirements.txt
Django==4.2.3
django-extensions
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-cors-headers